I started to think seriously about this application should be rewritten. It is difficult to make corrections because all things are descentralized and repeated in many places. I'm just make it work with Fluent Mapping because I know it and I use it, hope someone try to make the changes for the other mappings too.
Fix: There is a problem when a Table has a column with the same name as the table. This generates a Class with a property named as the same class, and this is not allowed. I managed the problem adding a suffix to the property, far now is static "Var" but is easily replaceable with a app config value, I will try to do it later.